fastapi_app/requirements.txt
fastapi
uvicorn
tensorflow
optuna
pillow
prefect
pandas
scikit-learn
pydantic
python-multipart
loguru
psutil